Working visualizations for a single stop.
Next steps making this cleaner and more colorful to help with visual distinctions and lengthening edges.

After that make the the URL(s) and keys configurable based on release version

After that incorporate the UI elements of station, date, and time so they cause the graph to update dynamically.

Finally harden the UI aganst bad data frm the front end and API and make the javascript cleaner.